+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/discordsrv/condition/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * DiscordSRV conditions.
+ */
+package org.betonquest.betonquest.compatibility.discordsrv.condition;