+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/discordsrv/objective/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * DiscordSRV objectives.
+ */
+package org.betonquest.betonquest.compatibility.discordsrv.objective;