+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `statistic` placeholder to get statistic values from the player
 - `redstonepower` condition to require a block at a specific location to be powered to a certain level
 - `hikari_pooling` config option to use HikariCP for database connections which are more stable
+- `discordsrv` integration added with a condition and an objective for account linking
 ### Changed
 - `lore` of `item` placeholder can now be used without number to get all lines
 - `chat` notify io now only sends messages during a conversation if `bypassInterceptor` is set to `true`